Separated currency selector as component
